Move the tutorial code under docs instead of in the src. Change the tests to have an autuse fixture in tests/tutorial/conftest.py to add docs/tutorial to sys.path.
