+++ tests/tutorial/test_factory.py
@@ -5,12 +5,12 @@ from wired import ServiceRegistry
 
 @pytest.fixture
 def registry():
-    from wired.samples.factory import setup
+    from factory.app import setup
     r: ServiceRegistry = setup()
     return r
 
 
 def test_greet_a_customer(registry):
-    from wired.samples.factory import greet_a_customer
+    from factory.app import greet_a_customer
     actual = greet_a_customer(registry)
     assert 'Hello !!' == actual
