+++ tests/tutorial/test_overrides.py
@@ -5,14 +5,14 @@ from wired import ServiceRegistry
 
 @pytest.fixture
 def settings():
-    from wired.samples.overrides import Settings
+    from overrides import Settings
     settings = Settings(punctuation='!!')
     return settings
 
 
 @pytest.fixture
 def registry(settings):
-    from wired.samples.overrides import app_bootstrap
+    from overrides import app_bootstrap
     r: ServiceRegistry = app_bootstrap(settings)
     return r
 
@@ -20,7 +20,7 @@ def registry(settings):
 def test_sample_interactions(registry):
     # Integration-style test
 
-    from wired.samples.overrides import sample_interactions
+    from overrides import sample_interactions
     greetings = sample_interactions(registry)
     assert 'Override Mary !!' == greetings[0]
     assert 'Bonjour Henri !!' == greetings[1]
