+++ tests/tutorial/test_requests_views.py
@@ -5,14 +5,14 @@ from wired import ServiceRegistry
 
 @pytest.fixture
 def settings():
-    from wired.samples.requests_views import Settings
+    from requests_views import Settings
     settings = Settings(punctuation='!!')
     return settings
 
 
 @pytest.fixture
 def registry(settings):
-    from wired.samples.requests_views import app_bootstrap
+    from requests_views import app_bootstrap
     r: ServiceRegistry = app_bootstrap(settings)
     return r
 
@@ -20,7 +20,7 @@ def registry(settings):
 def test_sample_interactions(registry):
     # Integration-style test
 
-    from wired.samples.requests_views import sample_interactions
+    from requests_views import sample_interactions
     greetings = sample_interactions(registry)
     assert 'Hello Mary !!' == greetings[0]
     assert 'Bonjour Henri !!' == greetings[1]
