+++ tests/tutorial/test_settings.py
@@ -5,19 +5,19 @@ from wired import ServiceRegistry
 
 @pytest.fixture
 def settings():
-    from wired.samples.settings import Settings
+    from settings.app import Settings
     settings = Settings(punctuation='!!')
     return settings
 
 
 @pytest.fixture
 def registry(settings):
-    from wired.samples.settings import setup
+    from settings.app import setup
     r: ServiceRegistry = setup(settings)
     return r
 
 
 def test_greet_a_customer(registry):
-    from wired.samples.settings import greet_a_customer
+    from settings.app import greet_a_customer
     actual = greet_a_customer(registry)
     assert 'Hello !!' == actual
